Created initial unit tests
