+++ src/main/controller.js
@@ -1,13 +1,11 @@
 const { ipcRenderer } = require('electron');
 const os = require('os');
-const fs = require('fs/promises');
 const path = require('path');
 
 const { execSync } = require('child_process');
 const fileIcon = require('file-icon');
 
-const { pathLocations, commonSuffix } = require('../../utils/pathLocations');
-const { fileRegex } = require('../../utils/fileRegex');
+const { findAppFiles } = require('./index');
 
 const dropZone = document.getElementById('drag-drop-zone');
 const dropZoneText = document.getElementById('drag-drop-zone-text');
@@ -21,8 +19,11 @@ const filesHeaderTitle = document.getElementById('files-header-title');
 const filesImage = '../../assets/img/files.svg';
 const addFileImage = '../../assets/img/add_files.svg';
 
-const scoreThreshold = 0.4;
 const mojaveDarwinMinVersion = '18.0.0';
+
+const isValidApp = (appPath) => path.extname(appPath) === '.app';
+
+const getAppIcon = (bundleId) => fileIcon.buffer(bundleId);
 
 const getSelectedFiles = () => [...document.querySelectorAll('input[name=checkbox]:checked')].map((item) => item.value);
 
@@ -69,143 +70,11 @@ async function getBundleIdentifier(appName) {
   return bundleId.substring(0, bundleId.length - 1);
 }
 
-async function getComputerName() {
-  const compName = await execSync('scutil --get ComputerName').toString();
-  console.log('compName', compName);
-  // remove empty space at end of string
-  return compName.substring(0, compName.length - 1);
-}
-
 function appNameFromPath(appPath) {
   const pathArr = appPath.split('/');
   const appNameWithExt = pathArr[pathArr.length - 1];
   // remove .app extension
   return appNameWithExt.replace('.app', '');
-}
-
-function createNameVariations(appName, bundleId) {
-  const appNameNorm = appName.toLowerCase().replace(' ', '');
-  const appNameWithoutDot = appNameNorm.toLowerCase().replace('.', '');
-  const appNameUnderscore = appName.toLowerCase().replace(' ', '_');
-  const appNameDash = appName.toLowerCase().replace(' ', '-');
-  const appNameDot = appName.toLowerCase().replace(' ', '.');
-
-  const bundleIdNorm = bundleId.toLowerCase().replace(' ', '');
-
-  return [
-    appNameNorm,
-    appNameWithoutDot,
-    appNameUnderscore,
-    appNameDash,
-    appNameDot,
-    bundleIdNorm,
-  ];
-}
-
-const normalizeString = (str, spacer = '') => str.toLowerCase().replace(/ /g, spacer);
-
-async function getFilePatternArray(appName, bundleId) {
-  const nameVariations = createNameVariations(appName, bundleId);
-  const appNameNorm = normalizeString(appName);
-  const bundleIdNorm = normalizeString(bundleId);
-
-  let patternArray = [...nameVariations];
-
-  const appNameComponents = appNameNorm.split('.');
-  if (appNameComponents) patternArray.push(appNameComponents[0]);
-
-  const bundleIdComponents = bundleIdNorm.split('.');
-  if (bundleIdComponents.length > 2 && bundleIdComponents[bundleIdComponents.length - 1].toLowerCase() === 'app') {
-    patternArray.push(`${bundleIdComponents.slice(0, bundleIdComponents.length - 1).join('.')}`);
-  }
-
-  const appWithSuffix = new Set([]);
-  commonSuffix.forEach((suffix) => nameVariations.forEach((nameVariation) => appWithSuffix.add(`${nameVariation}${suffix}`)));
-
-  patternArray = [...patternArray, [...appWithSuffix]];
-
-  return patternArray;
-}
-
-let compNameGlob;
-
-function stripString(file) {
-  let transformedString = file;
-  fileRegex.forEach((regex1) => {
-    transformedString = transformedString.replace(regex1, '');
-  });
-
-  const normCompName = normalizeString(compNameGlob, '-')
-    .replace(/\u2019/g, '')
-    .replace(/\(/g, '')
-    .replace(/\)/g, '');
-
-  transformedString = transformedString.replace(normCompName, '');
-  return transformedString;
-}
-
-function isPatternInFile(patterns, fileToCheck) {
-  return patterns.find((filePatten) => {
-    if (fileToCheck.includes(filePatten)) {
-      const strippedFile = stripString(fileToCheck);
-
-      let score = 0;
-      const indexOfString = strippedFile.indexOf(filePatten);
-      for (let i = 0; i < strippedFile.length; i += 1) {
-        if (i === indexOfString) {
-          i += indexOfString + filePatten.length;
-          score += filePatten.length;
-        }
-        if (strippedFile[parseInt(i, 10)] === '.') score += 0.5;
-        if (strippedFile[parseInt(i, 10)] === '_') score += 0.5;
-      }
-      if (score / strippedFile.length > scoreThreshold) {
-        return true;
-      }
-      return false;
-    }
-    return false;
-  });
-}
-
-async function getAppIcon(bundleId) {
-  const iconBuffer = await fileIcon.buffer(bundleId);
-  return iconBuffer;
-}
-
-async function findAppFiles(appName, bundleId) {
-  try {
-    compNameGlob = await getComputerName();
-    const bundleIdComponents = bundleId.split('.');
-
-    const companyDirs = pathLocations.map((pathLocation) => `${pathLocation}/${bundleIdComponents[1]}`);
-    const pathsToSearch = [...pathLocations, ...companyDirs];
-    const directoryFilesPromiseArr = pathsToSearch.map((pathLocation) => fs.readdir(pathLocation));
-    const directoryFiles = await Promise.allSettled(directoryFilesPromiseArr);
-
-    const patternArray = await getFilePatternArray(appName, bundleId);
-
-    const filesToRemove = new Set([]);
-
-    directoryFiles.forEach((dir, index) => {
-      if (dir.status === 'fulfilled') {
-        dir.value.forEach((dirFile) => {
-          const dirFileNorm = dirFile.toLowerCase().replace(' ', '');
-          if (
-            isPatternInFile(patternArray, dirFileNorm)
-          ) {
-            filesToRemove.add(`${pathsToSearch[parseInt(index, 10)]}/${dirFile}`);
-          }
-        });
-      }
-    });
-
-    // convert set to array
-    return [...filesToRemove];
-  } catch (err) {
-    console.error(err);
-    throw err;
-  }
 }
 
 function listItem(filePath, index) {
@@ -234,8 +103,6 @@ function listItem(filePath, index) {
   div.appendChild(label);
   fileList.appendChild(div);
 }
-
-const isValidApp = (appPath) => path.extname(appPath) === '.app';
 
 async function appSelectionHandler(appPath) {
   loadingContainer.style.display = 'flex';
@@ -269,6 +136,13 @@ async function appSelectionHandler(appPath) {
   loadingContainer.style.display = 'none';
 }
 
+async function openAppSelector() {
+  loadingContainer.style.display = 'flex';
+  const selectedApp = await ipcRenderer.invoke('selectAppFromFinder');
+  if (selectedApp) appSelectionHandler(selectedApp);
+  else loadingContainer.style.display = 'none';
+}
+
 deleteButton.addEventListener('click', async () => {
   const selectedFiles = getSelectedFiles();
 
@@ -286,13 +160,6 @@ deleteButton.addEventListener('click', async () => {
 clearButton.addEventListener('click', () => {
   clearList();
 });
-
-async function openAppSelector() {
-  loadingContainer.style.display = 'flex';
-  const selectedApp = await ipcRenderer.invoke('selectAppFromFinder');
-  if (selectedApp) appSelectionHandler(selectedApp);
-  else loadingContainer.style.display = 'none';
-}
 
 dropZone.addEventListener('click', openAppSelector);
