+++ src/main/index.html
@@ -45,6 +45,6 @@
     </div>
 
     <!-- You can also require other files to run in this process -->
-    <script src="./index.js"></script>
+    <script src="./controller.js"></script>
   </body>
 </html>
