+++ utils/funcs.js
@@ -1,0 +1,5 @@
+const isArrayUniqueValues = (arr) => Array.isArray(arr) && new Set(arr).size === arr.length;
+
+module.exports = {
+  isArrayUniqueValues,
+};
